Move info_utils errors to exceptions module
